mosquitto/requirements.txt
avro-python3==1.10.2
paho-mqtt==1.6.1